requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==43.0.1
Django==4.2.13
django-appconf==1.0.6
django-cryptography==1.1
django-environ==0.11.2
gunicorn==22.0.0
packaging==24.1
psycopg==3.1.19
psycopg2-binary==2.9.9
pycparser==2.22
sqlparse==0.5.0
typing_extensions==4.12.2